fix: fix critical bring-up bugs found during 3hr debug session - linker script offset math placed kernel at phys 0x400000   instead of 0x200000, updated linker script to match this address   along with stage 2 bootloader KERNEL_PHYS_ADDR - IDT gates stored 32-bit truncated handler addresses   (idt64_set_entry signature) - TSS RSP0/IST1 were garbage; moved stacks to .bss via linker script - HHDM offset applied through void* arithmetic causing silent truncation - pmm_alloc returned void* physical address enabling implicit truncation
several latent bugs would have caused silent corruption later.
squashed them all in one session.
never again.

+++ include/common/idt64.h
@@ -73,8 +73,8 @@ typedef void (*IDT64ISRDispatch)(IDT64ISRFrame *frame);
 #define IDT64_IDT_ATTR_TRAP       0x8F
 
 void idt64_load_idtr(IDT64Ptr *idt_ptr);
-void idt64_set_entry_int(IDT64Entry *idt_table, uint32_t index, uint32_t handler, uint8_t ist);
-void idt64_set_entry_trap(IDT64Entry *idt_table, uint32_t index, uint32_t handler, uint8_t ist);
+void idt64_set_entry_int(IDT64Entry *idt_table, uint32_t index, uint64_t handler, uint8_t ist);
+void idt64_set_entry_trap(IDT64Entry *idt_table, uint32_t index, uint64_t handler, uint8_t ist);
 void idt64_set_entries(IDT64Entry *entries, IDT64ISRHandler *handlers, size_t size);
 void idt64_enable_interrupts(void);
 void idt64_disable_interrupts(void);
